requirements.txt
altair==5.4.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
filelock==3.16.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
narwhals==1.13.2
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.3
pyarrow==18.0.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
setuptools==75.3.0
six==1.16.0
smmap==5.0.1
streamlit==1.39.0
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
toml==0.10.2
torch==2.5.1
tornado==6.4.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
ffmpeg-python
